backend/requirements.txt
fastapi
uvicorn
trafilatura
psycopg[binary]>=3.1
psycopg-pool>=3.1
litellm
pyjwt
scikit-learn
umap-learn
cohere
numpy
pydantic
markdown
requests
yt-dlp
PyMuPDF
pymupdf4llm
beautifulsoup4
youtube-transcript-api
google-api-python-client
google-auth
google-auth-oauthlib
langchain
langchain-openai
langgraph
